+++ test.mybricks.json
@@ -67,6 +67,7 @@
         "./src/form-coms/cascader/com.json",
         "./src/form-coms/treeSelect/com.json",
         "./src/form-coms/date-picker/com.json",
+        "./src/form-coms/date-picker-full-open/com.json",
         "./src/form-coms/range-picker/com.json",
         "./src/form-coms/time-picker/com.json",
         "./src/form-coms/time-range-picker/com.json",
